server/requirements.txt
fastapi[all]==0.99.1
pydantic==1.10.13
uvicorn==0.23.2
motor==3.6.0   # MongoDB async driver for FastAPI
aiohttp==3.8.6  # For asynchronous HTTP requests if required
sqlalchemy==2.0.29  # If using SQLAlchemy ORM in addition to MongoDB
dnspython==2.6.1   # For DNS-related utilities in your project
python-dotenv==1.0.0  # For environment variable management
pydantic[email]  # or email-validator
pymongo[srv]  # or pymongo[tls]
google-auth==2.36.0  # For Google OAuth2 authentication

# For development purposes
black==21.12b0
pytest==7.2.0
passlib==1.7.4
flake8==4.0.1
python-jose==3.3.0
